+++ src/core/src/utils/index.ts
@@ -2,7 +2,6 @@ export * from "./abort.js";
 export * from "./array.js";
 export * from "./async.js";
 export * from "./capability-probes.js";
-export * from "./date.js";
 export * from "./enumerated-options.js";
 export * from "./environment.js";
 export * from "./error.js";
